# Troller BK performance improvement by downloading files concurrently and make add to Solr async.