--- a/cl/corpus_importer/test_assets/import.csv
+++ b/cl/corpus_importer/test_assets/import.csv
@@ -1,0 +1,13 @@
+sources/troller-files/o-609|1575330086
+sources/troller-files/o-609|1575333374
+sources/troller-files/o-609|1575336978
+sources/troller-files/o-609|1575340576
+sources/troller-files/o-609|1575344176
+sources/troller-files/o-609|1575380176
+sources/troller-files/o-609|1575383907
+sources/troller-files/o-609|1575387437
+sources/troller-files/o-609|1575398235
+sources/troller-files/o-609|1575412625
+sources/troller-files/o-609|1575434178
+sources/troller-files/o-609|1576172179
+sources/troller-files/o-609|1582259778