This implements two of the performance improvements commented on in #2091.

- Moved the `add_items_to_solr` to celery, which took almost half of the processing time in large files with new entries, so this should help a lot.
- Download files concurrently and add them to a `queue` so that we can have some files ahead in memory so no download waiting time when there are files in the queue.

This was implemented using `Threads` and `Queue`, at the beginning the Queue is empty so `download_files_concurrently` starts by downloading `FILES_BUFFER_THRESHOLD` files in case there are less than this number of files in the queue (so the queue don't grow to the infinite, this number might be tweaked according to the behavior in production but in general, 3, I think it should work well).  If the queue is empty the process will wait until we have the first files in the queue but this waiting is faster than downloading 3 files in a single thread since this is concurrently using `concurrent.futures`.

In order to preserve the chronological order, downloaded files are ordered before adding to the Queue and it also waits for the previous thread to complete so the global order of files in the queue is not messed up by a thread in case their downloads are faster than others.

Once there are files in the `Queue` the regular process of parsing and merging the files starts since the `Queue` default order is `FIFO` we just pull out the file and then mark it as completed so it's removed from the Queue.


Here are some statics about this performance improvement related to download time:

Downloading 18 files of different sizes without the download queue took: 27.892451524734497 seconds
An average of 1.549555555555556 seconds per file.

Downloading the same 18 files concurrently and storing them in a queue took: 6.328908205032349 seconds
An average of  0.3516 seconds per file.

There are some times when the waiting is longer (approx. 2 seconds) when the queue is empty but in general, there is no waiting time when there is a file available in the queue.

- Added the `DataError` in the exception when saving a docket

Let me know what you think.